requirements.txt
streamlit
python-docx
docx2pdf
PyPDF2
pypandoc
watchdog
streamlit-option-menu
pyodbc
pandas